pyproject.toml
[tool.poetry]
name = "electronic-store"
version = "0.1.0"
description = ""
authors = ["Nikita-Musk <nik52119w@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
black = "^24.10.0"
isort = "^5.13.2"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
